+++ externals/coda-oss/modules/c++/logging/unittests/test_rotating_log.cpp
@@ -53,9 +53,9 @@ TEST_CASE(testRotate)
     {
         logging::Logger log("test");
 
-        auto logHandler = coda_oss::make_unique<logging::RotatingFileHandler>(outFile, 10, maxFiles);
+        auto logHandler = std::make_unique<logging::RotatingFileHandler>(outFile, 10, maxFiles);
         logHandler->setLevel(logging::LogLevel::LOG_DEBUG);
-        logHandler->setFormatter(coda_oss::make_unique<logging::StandardFormatter>("%m"));
+        logHandler->setFormatter(std::make_unique<logging::StandardFormatter>("%m"));
         log.addHandler(std::move(logHandler));
 
         log.debug("0123456789");