+++ externals/coda-oss/modules/c++/mem/unittests/test_unique_ptr.cpp
@@ -48,12 +48,12 @@ TEST_CASE(testStdUniquePtr)
         TEST_ASSERT_EQ(123, fooCtor->mVal);
     }
     {
-        auto fooCtor = coda_oss::make_unique<Foo>(123);
+        auto fooCtor = std::make_unique<Foo>(123);
         TEST_ASSERT_NOT_EQ(nullptr, fooCtor.get());
         TEST_ASSERT_EQ(123, fooCtor->mVal);
     }
     {
-        auto pFoos = coda_oss::make_unique<Foo[]>(123);  // 123 instances of Foo
+        auto pFoos = std::make_unique<Foo[]>(123);  // 123 instances of Foo
         TEST_ASSERT_NOT_EQ(nullptr, pFoos.get());
         TEST_ASSERT_EQ(0, pFoos[0].mVal);
         TEST_ASSERT_EQ(0, pFoos[122].mVal);
@@ -63,12 +63,12 @@ TEST_CASE(testStdUniquePtr)
 TEST_CASE(test_make_unique)
 {
     {
-        auto fooCtor = coda_oss::make_unique<Foo>(123);
+        auto fooCtor = std::make_unique<Foo>(123);
         TEST_ASSERT_NOT_EQ(nullptr, fooCtor.get());
         TEST_ASSERT_EQ(123, fooCtor->mVal);
     }
     {
-        auto pFoos = coda_oss::make_unique<Foo[]>(123);  // 123 instances of Foo
+        auto pFoos = std::make_unique<Foo[]>(123);  // 123 instances of Foo
         TEST_ASSERT_NOT_EQ(nullptr, pFoos.get());
         TEST_ASSERT_EQ(0, pFoos[0].mVal);
         TEST_ASSERT_EQ(0, pFoos[122].mVal);