+++ externals/coda-oss/modules/c++/sys/include/sys/ConditionVar.h
@@ -21,41 +21,28 @@
  */
 
 
-#ifndef __SYS_CONDITION_VAR_H__
-#define __SYS_CONDITION_VAR_H__
+#ifndef CODA_OSS_sys_ConditionVar_h_INCLUDED_
+#define CODA_OSS_sys_ConditionVar_h_INCLUDED_
+#pragma once
 
-#    if defined(USE_NSPR_THREADS)
-#        include "sys/ConditionVarNSPR.h"
-namespace sys
-{
-typedef ConditionVarNSPR ConditionVar;
-}
-#    elif (defined(WIN32) || defined(_WIN32))
-#        include "sys/ConditionVarWin32.h"
-namespace sys
-{
-typedef ConditionVarWin32 ConditionVar;
-}
-#    elif defined(__sun)
-#        include "sys/ConditionVarSolaris.h"
-namespace sys
-{
-typedef ConditionVarSolaris ConditionVar;
-}
-#    elif defined(__sgi)
-#        include "sys/ConditionVarIrix.h"
-namespace sys
-{
-typedef ConditionVarIrix ConditionVar;
-}
-//default to POSIX
-#    else
-#        include "sys/ConditionVarPosix.h"
-namespace sys
-{
-typedef ConditionVarPosix ConditionVar;
-}
-#    endif // Which thread package?
+#include "sys/Conf.h"
 
-#endif // End of header
+#if defined(_WIN32)
+#include "sys/ConditionVarWin32.h"
+namespace sys
+{
+using ConditionVar = ConditionVarWin32;
+}
 
+#elif defined(CODA_OSS_POSIX_SOURCE)
+#include "sys/ConditionVarPosix.h"
+namespace sys
+{
+using ConditionVar = ConditionVarPosix;
+}
+
+#else
+#error "Which thread package?"
+#endif
+
+#endif  // CODA_OSS_sys_ConditionVar_h_INCLUDED_