+++ externals/coda-oss/modules/c++/sys/unittests/test_conditionvar.cpp
@@ -30,28 +30,42 @@ typedef mt::CriticalSection<sys::Mutex> CriticalSection;
 TEST_CASE(testDefaultConstructor)
 {
     sys::ConditionVar cond;
+    TEST_SUCCESS;
 
     cond.acquireLock();
+    TEST_SUCCESS;
+
     cond.dropLock();
+    TEST_SUCCESS;
 }
 
 TEST_CASE(testParameterizedConstructor)
 {
     sys::Mutex mutex;
-    sys::ConditionVar cond(&mutex, false);
+    TEST_SUCCESS;
+
+    sys::ConditionVar cond(mutex);
+    TEST_SUCCESS;
 
     cond.acquireLock();
+    TEST_SUCCESS;
+
     cond.dropLock();
+    TEST_SUCCESS;
 }
 
 TEST_CASE(testMultipleTimeouts)
 {
     sys::Mutex mutex;
-    sys::ConditionVar cond(&mutex, false);
+    TEST_SUCCESS;
+
+    sys::ConditionVar cond(mutex);
+    TEST_SUCCESS;
 
     for (size_t ii = 0; ii < 5; ++ii)
     {
-        CriticalSection scopedLock(&mutex);
+        CriticalSection scopedLock(mutex);
+        TEST_SUCCESS;
         try
         {
             cond.wait(0.001);
@@ -61,8 +75,11 @@ TEST_CASE(testMultipleTimeouts)
         }
     }
     {
-        CriticalSection scopedLock(&mutex);
+        CriticalSection scopedLock(mutex);
+        TEST_SUCCESS;
+
         cond.broadcast();
+        TEST_SUCCESS;
     }
 }