+++ externals/nitro/modules/c++/nitf/include/nitf/UnitTests.hpp
@@ -33,7 +33,7 @@ namespace nitf
 {
 	namespace Test
 	{
-		NITRO_NITFCPP_API std::string buildPluginsDir();
+		NITRO_NITFCPP_API std::string buildPluginsDir(const std::string& dir = "nitf");
 		NITRO_NITFCPP_API std::filesystem::path buildFileDir(const std::filesystem::path& relativePath);
 		NITRO_NITFCPP_API std::filesystem::path findInputFile(const std::filesystem::path&);
 	}